Create uploading file
This code creates two functions, "makeCacheMatrix" and "cacheSolve". The "makeCacheMatrix" function creates a special "matrix" object that can cache its inverse, while the "cacheSolve" function computes the inverse of the special "matrix" returned by the "makeCacheMatrix" function.

The "cacheSolve" function first checks if the inverse has already been calculated and stored in the cache. If it has, it retrieves the inverse from the cache, and skips the computation. If not, it calculates the inverse of the matrix using the "solve" function in R, and then stores the inverse in the cache via the "setinverse" function.

The goal of this code is to provide a way to efficiently cache and retrieve the inverse of a matrix, so that the time-consuming inverse calculation does not have to be performed repeatedly